ai-task-system/v2.4/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiosqlite==0.19.0
pydantic==2.5.0
httpx==0.26.0
pytest==7.4.0
pytest-asyncio==0.23.0
openai>=1.12.0
websockets>=12.0
pyyaml>=6.0